simplify main.c
