+++ src/utils.h
@@ -1,0 +1,10 @@
+
+#ifndef PS2UTILS_H
+#define PS2UTILS_H
+
+#include <tamtypes.h>
+
+int print_buffer(qword_t *b, int len);
+int last_index_of(const char *str, int str_len, char c);
+
+#endif
